i965/drm: Make stride/pitch a uint32_t.
struct drm_i915_gem_set_tiling's stride field is a __u32.
intel_mipmap_tree::stride is a uint32_t.  Using unsigned long just
doesn't make sense.  Switching also lets us drop many pointless
locals that only existed to deal with the type mismatch.

+++ src/mesa/drivers/dri/i965/brw_bufmgr.h
@@ -109,7 +109,7 @@ struct brw_bo {
     */
    uint32_t tiling_mode;
    uint32_t swizzle_mode;
-   unsigned long stride;
+   uint32_t stride;
 
    time_t free_time;
 
@@ -161,7 +161,7 @@ struct brw_bo *brw_bo_alloc_tiled(struct brw_bufmgr *bufmgr,
                                   const char *name,
                                   int x, int y, int cpp,
                                   uint32_t tiling_mode,
-                                  unsigned long *pitch,
+                                  uint32_t *pitch,
                                   unsigned long flags);
 
 /** Takes a reference on a buffer object */
